plan_agent/pyproject.toml
[project]
name = "frame-summary-agent"
version = "0.1.0"
description = "A standalone agent for summarizing frame trades."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "a2a-sdk>=0.2.9",
]

[tool.setuptools]
py-modules = ["agent", "agent_executor"]
